docs: Convert memory.txt to rst format
Convert the memory API documentation from plain text
to restructured text format.

This is a very minimal conversion: all I had to change
was to mark up the ASCII art parts as Sphinx expects
for 'literal blocks' and to add a blank line for the
one example of a nested bulleted list.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>