tb-hash: improve tb_jmp_cache hash function in user mode
Optimizations to cross-page chaining and indirect jumps make
performance more sensitive to the hit rate of tb_jmp_cache.
The constraint of reserving some bits for the page number
lowers the achievable quality of the hashing function.

However, user-mode does not have this requirement. Thus,
with this change we use for user-mode a hashing function that
is both faster and of better quality than the previous one.

Measurements:

-    specINT 2006 (test set), x86_64-linux-user. Host: Intel i7-4790K @ 4.00GHz

XXX

-   specINT 2006 (train set), x86_64-linux-user. Host: Intel i7-4790K @ 4.00GHz

XXX

-               NBench, arm-linux-user. Host: Intel i7-4790K @ 4.00GHz

XXX

-              NBench, x86_64-linux-user. Host: Intel i7-4790K @ 4.00GHz

XXX

Signed-off-by: Emilio G. Cota <cota@braap.org>