tcg: bring parallel_cpus to tb->cflags and use it for TB hashing
This allows us to avoid flushing TB's when parallel_cpus is set.

Note that the declaration of parallel_cpus is brought to exec-all.h
to be able to define there the inlines. The inlines use an unnecessary
temp variable that is there just to make it easier to add more bits
to the mask in the future.

Signed-off-by: Emilio G. Cota <cota@braap.org>