qht: return existing entry when qht_insert fails
The meaning of "existing" is now changed to "matches in hash and
ht->cmp result". This is saner than just checking the pointer value.

Note that we do not just return void * (returning the inserted pointer
or the existing pointer) because it could be that the existing
entry is the same pointer we wanted to insert. So we still need another
return parameter (the bool) to know for sure that the insertion succeeded.

Signed-off-by: Emilio G. Cota <cota@braap.org>