trace: enable tracing of TCG atomics
These are memory accesses and therefore should be traced. Otherwise we get
different traces depending on whether we're in a single or parallel
environment.

Tested with a modified atomic_add-bench so that it is executes a deterministic
number of instructions, i.e. fixed seeding, no threading and fixed number of
loop iterations instead of running for a certain time.

Before:
- With parallel_cpus = false (no clone syscall so it is never set to true):
  220070 memory accesses
- With parallel_cpus = true (hard-coded):
  212105 memory accesses <-- we're not tracing the atomics!

After:
  220070 memory accesses regardless of parallel_cpus.

Signed-off-by: Emilio G. Cota <cota@braap.org>