thread: pass file and line to rec_mutex_lock/unlock/trylock
This paves the way for upcoming work that will need file/line
info on recursive mutexes.

Note that recursive mutexes are traced in Linux but not in
Windows. This commit doesn't fix that, because I don't have
a way to test changes on Windows. Besides, recursive mutexes
have very few callers so we might just not care enough about them.

Signed-off-by: Emilio G. Cota <cota@braap.org>