tests: add fp-test, a floating point test suite
This will allow us to run correctness tests against our
FP implementation. The test can be run in two modes (called
"testers"): host and soft. With the former we check the results
and FP flags on the host machine against the model.
With the latter we check QEMU's fpu primitives against the
model. Note that in soft mode we are not instantiating any
particular CPU (hence the HW_POISON_H hack to avoid macro poisoning);
for that we need to run the test in host mode under QEMU.

The input files are taken from IBM's FPGen test suite:
https://www.research.ibm.com/haifa/projects/verification/fpgen/

I see no license file in there so I am just downloading them
with wget. We might want to keep a copy on a qemu server though,
in case IBM takes those files down in the future.

The "IBM" syntax of those files (for now the only syntax supported
in fp-test) is documented here:
https://www.research.ibm.com/haifa/projects/verification/fpgen/papers/ieee-test-suite-v2.pdf

Note that the syntax document has some inaccuracies; the appended
parsing code works around some of those.

The appended also adds a few muladd-related operations that the IBM
syntax does not specify; model files for these are in muladd.fptest.

The exception flag (-e) is important: many of the optimizations
included in the following commits assume that the inexact flag
is set, so "-e x" is necessary in order to test those code paths.

The whitelist flag (-w) points to a file with test cases to be ignored.
I have put this file online, but we should have in on a QEMU-related server.

Thus, a typical of fp-test is as follows:

  $ cd qemu/build/tests/fp-test
  $ make -j && \
	./fp-test -t soft ibm/*.fptest \
	../../../tests/fp-test/muladd.fptest \
	-w ../../../tests/fp-test/whitelist.txt \
	-e x

The following commit fixes a possible bug in softfloat, but
after that is applied the above should return something like:

  All tests OK. Passed: 76532, not handled: 51237, whitelisted: 2748

I have successfully tested "host" mode on x86_64 and aarch64 machines.
On an IBM Power8 machine some exception flags are different from the
model files, which is interesting given that the model comes from IBM.

Running on host mode under QEMU reports flag mismatches (e.g. for
x86_64-linux-user), but that isn't too surprising given how little
love the i386 frontend gets. Host mode under aarch64-linux-user
passes OK.

Signed-off-by: Emilio G. Cota <cota@braap.org>