fpu: introduce hostfloat
The appended paves the way for leveraging the host FPU for a subset
of guest FP operations. For most guest workloads (e.g. FP flags
aren't ever cleared, inexact occurs often and rounding is set to the
default [to nearest]) this will yield sizable performance speedups.

The approach followed avoids checking the FP exception flags register.
See the comment at the top of hostfloat.c for details.

The licensing in softfloat.h seems quite complicated, so for simplicity
I'm adding this as a separate, GPL'ed file.

Signed-off-by: Emilio G. Cota <cota@braap.org>