util/threaded-workqueue: maximize batching
The round-robin policy is OK, but before moving to another
worker thread we better make sure this worker thread will be
busy for as long as possible.

Ideally the jobs we pass are as large as possible, so that
the overhead of communication (which can include sleeping)
is amortized. However, sometimes the jobs aren't that large,
and thus we are better off filling up each worker's work queue
before moving on to the next. Otherwise, with large numbers of
workers and small-ish jobs, we can end up underutilizing the
worker threads.

Signed-off-by: Emilio G. Cota <cota@braap.org>